Debug logging
This adds a logging feature which will hopefully be useful for diagnosing issues. A log file will be written to %TEMP%\FakeVive.log.

Additionally, you can pass a -FakeViveDebug command-line flag to the host program in order to show a console window and print extra log lines related to debugging.